requirements.txt
pillow
torch
torchvision
tqdm